+++ composer.json
@@ -17,7 +17,7 @@
         "symfony/filesystem": "^4.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.5",
+        "phpunit/phpunit": "^7.3",
         "mikey179/vfsStream": "~1.4.0",
         "symfony/config": "^4.1",
         "symfony/dependency-injection": "^4.1",
@@ -26,5 +26,8 @@
     },
     "autoload": {
         "psr-4": { "Beryllium\\Icelus\\": "" }
+    },
+    "autoload-dev": {
+        "psr-4": { "Beryllium\\Icelus\\": "tests/Beryllium/Icelus" }
     }
 }
