+++ phpunit.xml.dist
@@ -7,8 +7,7 @@
         convertNoticesToExceptions="true"
         convertWarningsToExceptions="true"
         processIsolation="false"
-        stopOnFailure="false"
-        syntaxCheck="false">
+        stopOnFailure="false">
     <testsuites>
         <testsuite name="Beryllium/Icelus">
             <directory>./tests/</directory>
