+++ tests/Beryllium/Icelus/ImageServiceTest.php
@@ -3,41 +3,16 @@
 namespace Beryllium\Icelus;
 
 use Imanee\Exception\ImageNotFoundException;
-use Imanee\Imanee;
-use org\bovigo\vfs\vfsStream;
-use org\bovigo\vfs\vfsStreamDirectory;
 use Symfony\Component\Filesystem\Filesystem;
 
-class ImageServiceTest extends \PHPUnit_Framework_TestCase
+class ImageServiceTest extends IcelusTestBase
 {
-    /**
-     * @var string
-     */
-    public $source_dir;
-
-    /**
-     * @var Imanee
-     */
-    public $imanee;
-
-    /**
-     * @var vfsStreamDirectory
-     */
-    public $output_dir;
-
-    public function setUp()
-    {
-        $this->imanee     = new Imanee;
-        $this->source_dir = __DIR__ . '/../../Resources';
-        $this->output_dir = vfsStream::setup('thumbs');
-    }
-
     public function testValidThumbnail()
     {
         $service = new ImageService(
             $this->imanee,
             $this->source_dir,
-            $this->output_dir->url('thumbs'),
+            $this->output_writer,
             null,
             new Filesystem
         );
@@ -52,7 +27,7 @@ class ImageServiceTest extends \PHPUnit_Framework_TestCase
         $service = new ImageService(
             $this->imanee,
             $this->source_dir,
-            $this->output_dir->url('thumbs'),
+            $this->output_writer,
             null,
             new Filesystem
         );
@@ -67,11 +42,13 @@ class ImageServiceTest extends \PHPUnit_Framework_TestCase
 
     public function testOutputDirNotFound()
     {
-        $dir = $this->output_dir->url('thumbs');
+        $writer = clone $this->output_writer;
+        $writer->setOutputDir($writer->getOutputDir() . '/test');
+
         $service = new ImageService(
             $this->imanee,
             $this->source_dir,
-            $dir . '/test',
+            $writer,
             null,
             new Filesystem
         );
