+++ tests/Beryllium/Icelus/TwigImageExtensionTest.php
@@ -3,43 +3,13 @@
 namespace Beryllium\Icelus;
 
 use Imanee\Exception\ImageNotFoundException;
-use Imanee\Imanee;
-use org\bovigo\vfs\vfsStream;
-use org\bovigo\vfs\vfsStreamDirectory;
 use Symfony\Component\Filesystem\Filesystem;
 
-class TwigImageExtensionTest extends \PHPUnit_Framework_TestCase
+class TwigImageExtensionTest extends IcelusTestBase
 {
-    /**
-     * @var string
-     */
-    public $source_dir;
-
-    /**
-     * @var Imanee
-     */
-    public $imanee;
-
-    /**
-     * @var vfsStreamDirectory
-     */
-    public $output_dir;
-
-    /**
-     * @var Twig
-     */
-    public $twig;
-
-    public function setUp()
-    {
-        $this->imanee     = new Imanee;
-        $this->source_dir = __DIR__ . '/../../Resources';
-        $this->output_dir = vfsStream::setup('thumbs');
-    }
-
     public function testValidThumbnail()
     {
-        $service = new ImageService($this->imanee, $this->source_dir, $this->output_dir->url('thumbs'), null, new Filesystem);
+        $service = new ImageService($this->imanee, $this->source_dir, $this->output_writer, null, new Filesystem);
 
         // test a valid resource
         $thumbnail = $service->thumbnail('valid.jpg', 100, 100, false);
@@ -48,7 +18,7 @@ class TwigImageExtensionTest extends \PHPUnit_Framework_TestCase
 
     public function testNotFoundThumbnail()
     {
-        $service = new ImageService($this->imanee, $this->source_dir, $this->output_dir->url('thumbs'), null, new Filesystem);
+        $service = new ImageService($this->imanee, $this->source_dir, $this->output_writer, null, new Filesystem);
 
         // test a not-found resource
         try {
